feat(l2): multi prover foundations (#1397)
**Motivation**

We want to have multiple `prover_clients` with different proving systems
running in parallel, then the `prover_server` should keep track of each
proving system and their proofs to verify them on chain.

When all the required proving systems finish to create the blocks proof,
and then verified on chain, we can say that the blockchain is advancing.

At the moment, just one proving system is allowed, there isn't a
mechanism to keep track of many `prover_clients` with different proving
systems.

**Description**

- Introduce a common API to access different proving systems.
- Deployment and modification of contracts.
- Switch between the `prover_client's` proving systems