feat(levm): add new CI directive to show tests result summary (#1557)
**Motivation**

In LEVM, every time we need to check how many tests pass, we need to
open the `Run tests` directive and scroll all the day down where the
Summary section is presented.
With this patch, a new CI directive called `Show test summary` is
created, where only the Summary is shown; thus saving some time.

**Description**

When the CI runs the `make run-evm-ef-tests-ci` directive, it is piped
to `tee` which outputs the results to stdout and also a file.
Then, the `Show test summary` directive simply `awk`s said file for the
summary.