fix(levm): sstore refund logic (#1559)
**Motivation**

There was a discrepancies in gas usage and refunds. The issue was
related to an improper management of gas refunds in `sstore`.

**Description**

There was an error when calculating the gas refund. Adjust the
conditional logic to subtract refunds when reverting storage changes and
adding refunds appropriately. Also, reordered refund and cost logic.