fix(levm): remove accounts in selfdestruct (#1566)
**Motivation**

We were not handling the remove case well in `selfdestruct`. When
removing it directly from from the cache, it generated a discrepancy in
the root.

**Description**

- When we have addresses in `accrued_substate.selfdestruct_set`, we
assign the default value to the account, marking the account as empty.
- We check in `get_state_transitions` if `new_state_account.is_empty()`
and assign the value to the `removed` field.
- Handle the case when `target_account` is the same as
`current_call_frame.to`. In that case, we should burn the eth.