feat(levm): run hive tests with levm, generate and publish report (#1546)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- We now can execute multiple transactions in a block and commit them
after execution of them all.
- It refactors `get_state_transitions_levm`, now in `account_updates`
we'll have only the things that have changed from the account, so that
we don't have much redundant information

- We build the client image with LEVM, we run hive tests and then we
publish them
- Add to the Dockerfile the possibility to add build flags.
- Avoid panicking if setup wasn't completed successfully, as no tests
will run in that scenario.


We can run tests with hive but the problem is that these tests execute a
lot of blocks at the beginning and we usually fail at some block because
of a difference in state, I tried debugging some tests and I fixed some
things in other PRs but at some point I got stuck and decided that it
was a better idea to leave it as is and merge this to main.