feat(levm): support Shanghai version (#1600)
**Motivation**

The idea is to support the Shanghai version.

**Description**

- Remove the address `0x10` (`point_evaluation` precompile) since it
does not exist in pre-Cancun.
- Return `Type3TxPreFork` if `tx_max_fee_per_blob_gas` is Some and we
are in pre-Cancun.
- There was a problem in `levm_runner.rs/get_state_transitions()` when
attempting to self-destruct an account by sending a balance of 0 to an
existing account with no nonce and no balance. The previous code
interpreted that there was no update in the new state, so the account
was not added to the account update list. We need to mark this account
as removed.

`Shanghai: 221/221 (100.00%)`

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>