feat(l1): show disconnect reason meaning when a peer disconnects (#1618)
**Motivation**
When a peer disconnects there is an info! that shows the disconnect
error code, but this is not very helpful. This PR replaces this error
code with the corresponding error meaning according to the spec.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add method `reason` for rlpx Disconnect message which returns the
meaning of the reason code as a str and uses it when tracing disconnects
from peers in their listen loop.
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None, this is a minor improvement