fix(l1): fix hive LargeTxRequest test (#1480)
The fix consists in removing the transaction limit in the Transactions
rlpx decode. While there might be a limit to transaction responses, we
should not limit incoming transactions and specially not in the decoding
stage, as this is an eth wire protocol limit, not an rlpx limit.

Additional changes: added error messages and more detailed disconnect
messages. For this PR this only helped to see that there were "early
eofs", which can roughly be transalted as abrupt disconnections from the
peer.