feat(l1): show the disconnect reason in the HandshakeError when a peer disconnects mid-handshake (#1667)
**Motivation**
When connecting to the sepolia testnet I encountered a lot of handshake
errors but all of them said "Expected Hello" or "Expected Status" with
no other information. Most of these handshake failures are due to the
peer sending a disconnect message instead. This PR aims to improve
observability & debugging experience by showing the disconnect reason in
the `HandshakeError`.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Show the disconnect reason in the `HandshakeError` message when a
`Disconnect` message is received during the handshake process (p2p)
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None