fix(levm): for pre Berlin don't use floor cost modexp (#1675)
**Motivation**

This PR is for fixing the Ethereum Foundation Tests on the ModExp
precompile.

**Description**

When the `base_size` and `modulus_size` is Zero, there is an early
return to not make all the operations.
In the Berlin fork was introduced a floor for the gas cost. We were
returning this floor cost also for the previous forks to Berlin.
Now is only added after Berlin.