feat(levm): increase calldata cost - EIP-7623 (#1665)
**Motivation**

This Pull Request exists to make a floor cost for the gas usage in the
transactions. This is specified in
[EIP-7623](https://eips.ethereum.org/EIPS/eip-7623), and aims to reduce
the maximum block size to make room for additional blobs or potential
block gas limit increases. Also, it introduces that the gas limit has to
be at least the floor cost.

**Description**

Here, for the Prague Spec adds the check in the `prepare_execution` to
the gas limit be greater or equal to the max value of the intrinsic gas
or `TX_BASE_COST + TOTAL_COST_FLOOR_PER_TOKEN * tokens_in_calldata`.
Where `TOTAL_COST_FLOOR_PER_TOKEN` is a new constant with the value of
10.
A new error is added when the gas limit is too low,
`TxValidationError::GasLimitTooLow`.

Then when creating the report, the gas usage for Prague is the max value
of the gas usage or `TX_BASE_COST + TOTAL_COST_FLOOR_PER_TOKEN *
tokens_in_calldata`.

**State**

The only tests for this EIP that fail are the test with type 4
transactions. This is because we don't have implemented them.

Closes #1640

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>