fix(l1): prevents HelloMessage decode from failing if there are unknown capabilities (#1671)
This was solved by adding an `UnsupportedCapability` in the Capability
enum. The rationale behind this is:

- We can't just represent all possible capabilities because some nodes
use off-spec capabilities and may add them whenever they want,
which would prevent us from connecting with them.
- We may represent them as RlpxDecode errors, but as we decode a vector
of them, the vector decoding function would need to be rewritten to not
break in this very specific case.
- If we represent this as a type, we don't have ramifications, as we
only use it to check if there's at least one of the other node
capabilities that we support.

Closes #1656