feat(levm): add precompile bls12_g1add (#1688)
**Motivation**

In this PR we implement the first precompile contract stated in the
[EIP-2537](https://eips.ethereum.org/EIPS/eip-2537) for the Prague fork

**Description**

For the address 0x0b, call the function bls12_g1add.
Parse the input data as to G1 Point of 128 bytes. A G1 Point consists of
two 64 bytes points, the first 16 bytes are padding of zeroes.
All this checks are made, if any fails, an error is raised.

Using the crate bls12_381, we create to G1 points from the raw data of
the call. Then the method add is called and the result is returned as
Bytes.

If any of the points of G1 is not in the curve nor is a point to
infinity, an error is raised.

Closes #1694

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>