feat(levm): add precompile bls12_g2add (#1717)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
In this PR we implement the third precompile contract stated in
[EIP-2537](https://eips.ethereum.org/EIPS/eip-2537) for the Prague fork,
the `BLS12_G2ADD`

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

For the address 0x0d, the function `bls12_g2add` is called.
First parse the input as two G2 points. Each point consists of four Fp
coordinates, each of 64 bytes. Then, the input must be 512 bytes long.
If the point parsed to a G2 is not on the curve nor a point to infinity,
an Error is raised.

Then the crate
[BLS12_381](https://github.com/sp1-patches/bls12_381/tree/main) is
called to perform the addition.
Note here: this crate serialize the data in a different order as the
input is sent, this is handled.

<!-- Link to issues: Resolves #111, Resolves #222 -->

All tests are passing: bls12_g2add: 46/46 (100.00%)

Closes #1696

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>