fix(l1): kademlia table lock & add peer to table on rlpx receiver connection (#1725)
**Motivation**
This PR started as a fix for the missing ancestor’s Hive tests. While it
doesn’t fix the tests, it finds two bugs that advance on the issue and
are worth merging asap as other tests might be failing for the same
reason.


**Description**
This PR required a handful of debugging, most of the initial causes for
the test failing were not correct:
- We missed syncing when getting nodes we didn't have their parent ->
This was already implemented
- Bootnodes weren't set so maybe the order of validation was wrong ->
Same validation on other test worked and most of them dependen on having
their ancestors
- We were missing some configuration on hive level -> We didn't
- We may be erroneously setting the http.addr -> It worked as other
clients

Near the second to third check, incoming peers appeared to be involved
in the issue and after a bunch of tests we could pinpoint incoming
connections actually fulfilling the handshake but failing afterward
without peers. This shown that while the p2p connection was made, the
peer wasn't persisted in discv4 kademblia table. This was solved by
adding the receiver connection via `tcp` to the `Kademlia` table.
However, the issue persisted because of the `get_peer_channels`
function, where we indefinitely locked the Kademlia table until a peer
was found. If no peers were in the table beforehand, this caused an
infinite loop. This situation occurred when there were no existing
peers, and we attempted to sync via the Snap protocol.

**TL;DR**
1.  Fixed a deadlock in the `Kademlia` table in `get_peer_channels`.
2. When receiving connections via `rlpx` we add the peer to the Kademlia
table.


Closes None

---------

Co-authored-by: Marcos Nicolau <marcosnicolau@lambdaclass.com>
Co-authored-by: Marcos Nicolau <76252340+MarcosNicolau@users.noreply.github.com>