fix(levm): remove unstable tests (#1767)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

This PR removes a folder of tests for the EIP-2537. This tests hasn't
been updated for 9 months.
400 tests were successfully passing with the existing implementation, no
modifications were required to address the issues.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

A
[folder](https://github.com/ethereum/tests/tree/develop/EIPTests/StateTests/stEIP2537)
containing 1403 tests for the precompiles added for Prague fork was
removed.
Removes the line from download and create a directory for this tests in
the Makefile.

<!-- Link to issues: Resolves #111, Resolves #222 -->