fix(levm): get correct calldata in creation contracts (#1784)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

Due to the changes introduced by EIP-7623, a new floor cost for gas was
introduced and it is calculated with the calldata. The problem arise
when there is a creation contract, this calldata is emptied and the
bytecode is needed to calculate the floor cost.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

In the function `gas_used`, instead of using the
`initial_call_frame.calldata`, if it is a create transaction,
`initial_call_frame.bytecode` is used.
A new variable `calldata` is introduced to store it.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number