chore(levm): ci: compare against new benchmarks in main (#1834)
**Motivation**

Since the recently #1803 PR we now have new benchmarks. We need to
modify the CI to make the comparison.
I couldn't do that in the previous PR because since the new benchmarks
were not in main yet, the CI would always fail.

**Description**
I will fix this in the next benchmark PR so we don't have to create 2
PRs for each new benchmark.
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->