refactor(levm): generalize the incrementation of the pc (#1827)
**Motivation**

Make the increment of the PC "generalized". Right now, we increment the
PC by 1 every time; which requires some additional logic for the cases
where the PC should _not_ be increment (such as JUMP and JUMPI).

**Description**

Every OPCODE now returns how much they should increment the PC. This
avoids issues where some Opcode increment the PC by different amount at
different times; thus leaving the window open for state changes.

With this the increment of the PC is done in a single place. 

Also, remove the following:
- Method `increment_pc` from VM.
- `pc()` method from VM.
- Remove `HaltReason` enum. It was not being used.
- `created_address` member variable from `TransactionReport` since it
was not being used.

Closes #1833