feat(l1): implement EIP-7840 (#1810)
**Motivation**

The execution client has to support the [Prague
update](https://eips.ethereum.org/EIPS/eip-7600)

**Description**

This PR implements the
[EIP-7840](https://eips.ethereum.org/EIPS/eip-7840) by adding new
configuration parameters to the `ChainConfig` struct and modifying the
usage of certain blob related constants in the code to make use of them.
If not configured in the genesis file, this parameter will be set to the
default specified in the EIP, which is:

```json
"blobSchedule": {
  "cancun": {
    "target": 3,
    "max": 6,
    "baseFeeUpdateFraction": 3338477
  },
  "prague": {
    "target": 6,
    "max": 9,
    "baseFeeUpdateFraction": 5007716
  }
}
```

This PR only modifies usage of said constants in the L1 code. Same
should be done in the LEVM and L2 crates, which originated the issues
#1813 #1814

Some issues were created while developing this PR:

First, the issue #1815 was created, so the different `constants.rs`
files are merged into just one `constant.rs` file under the root of
`common` crate. This file was already created in this PR due to the need
of having some blob related constants in the `ChainConfig` struct.

Second, the issue #1840 to check if the calculation of `GetBlobBaseFee`
is correctly made and correct it if it's not.

Third, the issue #1846 to add validations to Prague blocks on
`blockchain.rs` module.

---------

Co-authored-by: Julian Ventura <julian.ventura@lambdaclass.com>
Co-authored-by: Avila Gastón <72628438+avilagaston9@users.noreply.github.com>