fix(l1): log transaction errors instead of disconnecting (#1863)
**Motivation**

Currently, if we fail to process transactions received from peers, we
propagate the error until we disconnect from them. This is incorrect
behavior because we might be one block behind the current head and
invalidating transactions for insufficient balance.

**Description**

- Stops disconnecting from peers on validation errors. #1806 was
created. This change brakes the `BlobViolations` Hive test, which was a
false positive. We have #1781 to fix this.
- When receiving a `Transactions` message, filters and propagates only
the valid transactions. This change brakes the Hive test
`eth/InvalidTxs`, which was also a false positive since we are
incorrectly considering valid two transactions in the test. #1139 was
created to fix this.