feat(levm): add support for EIP-7840 and add `EVMConfig` to LEVM. (#1860)
**Motivation**

PR #1810 introduced support for EIP-7840 in L1. That EIP introduced the
notion that certain blobhash related constants could be modified by
Genesis file (for more information see commit:
8610b4cc87d25dab2b75574df94079a953aa6161).

**Description**

This introduces a new struct called `EVMConfig` to LEVM. It only holds
the current `Fork` and the desired `ForkBlobSchedule`.

In most cases, the default `ForkBlobSchedule` will want to be used (for
that purpose, the `EVMConfig::canonical_values(fork: Fork) ->
ForkBlobSchedule` functions exists). This is **unless** you want to use
a custom `ForkBlobSchedule` like described in
[EIP-7840](https://eips.ethereum.org/EIPS/eip-7840).


Closes #1813

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>