feat(l1): improve healing (#1867)
**Motivation**
Reverts the "slow but safe" approach to healing taken by #1763 and
instead adds the intermediate nodes received from peers to the trie.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Write nodes fetched during healing to the state instead of writing
only leaf values
* Track paths left in queue when a healing cycle ends due to pivot
staleness
* Condense snap checkpoint clearing to one single method
* (Misc) Move some noisy p2p info tracing to debug
* (Fix) When executing full blocks after the pivot block during snap
sync execute the block first before setting it as canonical
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->
**Comments**
A not so pretty solution was implemented on the storage_healer (the
startup flag) to avoid blocking storage healing at the start. This can
and will be improved but doing so in this PR is not worth it when follow
up PRs will already touch up on this logic when parallelizing storage
node fetching (similar to storage ranges)

**Testing**
This proved to work fine in `Mekong` testnet when introducing forced
sleeps and delays to force healing. But it is still too slow to work on
Holesky testnet, this should improve once we add more parallelization to
state sync & healing

Closes #issue_number