feat(l1): update EF tests to use Prague pre-release (#1838)
**Motivation**

We need a suite of tests to check our Prague implementation.

**Description**

This PR updates the EF tests to the latest [pre-release
version](https://github.com/ethereum/execution-spec-tests/releases/tag/pectra-devnet-5%40v1.3.0),
which includes Prague tests.

> [!IMPORTANT]
> Changes were introduced on how tests are parsed, so many tests will be
shown as passing. These are false positives.
> Read below

The EF tests are organized in different directories. There is one
directory per fork, each containing subdirectories for their implemented
EIPs. For each of said EIPs there are many json files, which contain the
information to run the tests. Those json files not only test the
corresponding EIP fork, but also the remaining ones. For example, under
`cancun` directory we may find some json files which not only test the
EIPs on a `cancun` fork, but also on `Prague`. This is also true for the
remaining forks.
Because of this, a filter was included in this PR so not-supported or
not-yet-implemented forks tests are ignored. The idea is to loose this
filter restriction while implementing Prague functionality. The crate
that it's being used for parsing these tests is `datatest_stable`, and
it does not provide a way of indicating that a test has been ignored, so
all ignored tests will be marked as passing, producing false positives.

---------

Co-authored-by: Julian Ventura <julian.ventura@lambdaclass.com>