chore(core): add evm flag to ethrex (#1884)
**Motivation**

We are adding some changes related to features flags in PR
https://github.com/lambdaclass/ethrex/pull/1861 . The `ci_levm.yaml`
performs a comparison between the PR's commit and main. To fully test
this JOB and based on this
[comment](https://github.com/lambdaclass/ethrex/pull/1861#discussion_r1945064630),
we should add the cli argument and do nothing for now, just to run the
tests with the latest commit from Hive.


**Description**

- Add the `evm` cli argument to avoid errors.
- Bump up Hive's commit.

Linked to: https://github.com/lambdaclass/ethrex/pull/1861